Backend/requirements.txt
django 
djangorestframework 
psycopg2-binary 
django-cors-headers 
python-dotenv 
pillow
django-cors-headers